% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_plot.R
\name{emm_plot_recording}
\alias{emm_plot_recording}
\title{Plot results from linear mixed effects model with emmeans}
\usage{
emm_plot_recording(model, df, y_var)
}
\arguments{
\item{model}{linear mixed effects model}

\item{df}{Dataframe used to fit the model}
}
\value{
plot with x axis as recording or ultrasound, y axis as response variable
}
\description{
Plot results from linear mixed effects model with emmeans
}
