% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_plot.R
\name{emm_plot_ultrasound}
\alias{emm_plot_ultrasound}
\title{Plot results from linear mixed effects model (ultrasound on x-axis)}
\usage{
emm_plot_ultrasound(model, df, y_var)
}
\arguments{
\item{model}{linear mixed effects model}

\item{df}{Dataframe}

\item{y_var}{column to plot}
}
\value{
ggplot object
}
\description{
Plot results from linear mixed effects model (ultrasound on x-axis)
}
