% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{tidydata}
\alias{tidydata}
\title{Tidy Excel data containing AQUA2 Max Dff}
\usage{
tidydata(df, timeinsec)
}
\arguments{
\item{df}{Dataframe needed to be tidied}

\item{timeinsec}{Time in seconds for each recording}
}
\value{
with cols for exp, coverslip, ultrasound, treatment, dFF
}
\description{
Tidy Excel data containing AQUA2 Max Dff
}
