% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_transform.R
\name{hz_transform}
\alias{hz_transform}
\title{If Hz histogram is not normal, this use Boxcox transformation}
\usage{
hz_transform(df, col_name)
}
\arguments{
\item{df}{tidy_df}

\item{col_name}{name of column to transform}
}
\value{
df with transformed Hz column
}
\description{
If Hz histogram is not normal, this use Boxcox transformation
}
